backend/requirements.txt
# Production dependencies
fastapi==0.128.4
uvicorn[standard]==0.40.0
pydantic==2.10.6
pydantic-settings==2.7.2
python-dotenv==1.0.1
pandas==3.0.0
networkx==3.6.1
openai==1.59.7
python-multipart==0.0.20
